Navwalker Added to WP
Added bootstrap navwalker.php to search for pages created in WP and allow them to appear on top menu. Wrote some jQuery to add classes necessary to producing the "bootstrap" style dropdown. Some fine tuning is still needed. Set the "body" color to the white that we chose in the color scheme.