Created footer.php
Added a widget area for donations as well as a log-in form for registered users. Changed the break-point of the "cards" to break at tablet, looks much better. Added more Fontawesome icons to design. Created (overwrote) "alert-secondary" which is being used in the footer. Notice part of the footer comes from the dashboard and the other part is in the raw code?